+++ database/sql/import-sql.sh
@@ -18,6 +18,8 @@
 #
 # Let's begin...
 
+source /srv/provision/provision-helpers.sh
+
 VVV_CONFIG=/srv/config/default-config.yml
 if [[ -f /srv/config/config.yml ]]; then
 	VVV_CONFIG=/srv/config/config.yml
@@ -27,12 +29,12 @@ run_restore=$(shyaml get-value general.db_restore 2> /dev/null < ${VVV_CONFIG})
 
 if [[ $run_restore == "False" ]]
 then
-	echo " * Skipping DB import script, disabled via the VVV config file"
+	vvv_info " * Skipping DB import script, disabled via the VVV config file"
 	exit;
 fi
 
 # Move into the newly mapped backups directory, where mysqldump(ed) SQL files are stored
-echo " * Starting MariaDB Database Import"
+vvv_info " * Starting MariaDB Database Import"
 # create the backup folder if it doesn't exist
 mkdir -p /srv/database/backups
 cd /srv/database/backups/
@@ -56,27 +58,28 @@ then
 		[ "${db_name}" == "performance_schema" ] && continue;
 		[ "${db_name}" == "test" ] && continue;
 
-		echo " * Creating the \`${db_name}\` database if it doesn't already exist, and granting the wp user access"
+		vvv_info " * Creating the <b>${db_name}</b><info> database if it doesn't already exist, and granting the wp user access"
 		mysql -u root --password=root -e "CREATE DATABASE IF NOT EXISTS \`${db_name}\`"
 		mysql -u root --password=root -e "GRANT ALL PRIVILEGES ON \`${db_name}\`.* TO wp@localhost IDENTIFIED BY 'wp';"
+
+		[ "${db_name}" == "wordpress_unit_tests" ] && continue;
 
 		mysql_cmd="SHOW TABLES FROM \`${db_name}\`" # Required to support hyphens in database names
 		db_exist=$(mysql -u root -proot --skip-column-names -e "${mysql_cmd}")
 		if [ "$?" != "0" ]
 		then
-			echo " * Error - Create the '${db_name}' database via init-custom.sql before attempting import"
+			vvv_error " * Error - Create the <b>${db_name}</b><error> database via init-custom.sql before attempting import"
 		else
-			if [ "" == "${db_exist}" ]
-			then
-				echo " * mysql -u root -proot '${db_name}' < '${db_name}.sql'"
+			if [ "" == "${db_exist}" ]; then
+				vvv_info " * Importing <b>${db_name}</b><info> from <b>${db_name}.sql</b>"
 				mysql -u root -proot "${db_name}" < "${db_name}.sql"
-				echo " * Import of '${db_name}' successful"
+				vvv_success " * Import of <b>'${db_name}'</b><success> successful</success>"
 			else
-				echo " * Skipped import of \`${db_name}\` - tables already exist"
+				vvv_info " * Skipped import of <b>\`${db_name}\`</b><info> - tables already exist"
 			fi
 		fi
 	done
-	echo " * Databases imported"
+	vvv_success " * Databases imported"
 else
-	echo " * No custom databases to import"
+	vvv_success " * No custom databases to import"
 fi