+++ provision/core/nodejs/provision.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -eo pipefail
 
 function nodejs_register_packages() {
   if ! vvv_src_list_has "nodesource"; then
@@ -7,7 +8,7 @@ function nodejs_register_packages() {
 
   if ! vvv_apt_keys_has 'NodeSource'; then
     # Retrieve the NodeJS signing key from nodesource.com
-    echo " * Applying NodeSource NodeJS signing key..."
+    vvv_info " * Applying NodeSource NodeJS signing key..."
     apt-key add /srv/config/apt-keys/nodesource.gpg.key
   fi
 
@@ -21,7 +22,7 @@ vvv_add_hook before_packages nodejs_register_packages
 
 function node_setup() {
   if [[ $(nodejs -v | sed -ne 's/[^0-9]*\(\([0-9]\.\)\{0,4\}[0-9][^.]\).*/\1/p') != '10' ]]; then
-    echo " * Downgrading to Node v10."
+    vvv_info " * Downgrading to Node v10."
     apt remove nodejs -y
     apt install -y --allow-downgrades --allow-remove-essential --allow-change-held-packages -o Dpkg::Options::=--force-confdef -o Dpkg::Options::=--force-confnew install --fix-missing --fix-broken nodejs
   fi
@@ -29,9 +30,9 @@ function node_setup() {
   # npm
   #
   # Make sure we have the latest npm version and the update checker module
-  echo " * Installing/updating npm..."
+  vvv_info " * Installing/updating npm..."
   npm_config_loglevel=error npm install -g npm
-  echo " * Installing/updating npm-check-updates..."
+  vvv_info " * Installing/updating npm-check-updates..."
   npm_config_loglevel=error npm install -g npm-check-updates
 }
 export -f node_setup